Added schedule pickup logic
Added the logic for selecting pickup date and pickup time.
As of now the values are hardcoded but plan to move into database or backend so that it is easy to update without updating app.
If the date is today and time it after the start of last time slot then user can not choose today for pickup.
If the date is today and time is overlapping any time slot then use cannot use that time slot or before that. so added the logic to disable the time slots accordingly